Moving settings sources to new home
