+++ UpdateArc/helpers/includes/settings.h
