+++ UpdateArc/helpers/sources/settings.cpp
