+++ UpdateArc/whereisit.pri
@@ -2,6 +2,7 @@ SOURCES += \
     helpers/sources/downloader.cpp \
     helpers/sources/logger.cpp \
     helpers/sources/version.cpp \
+    helpers/sources/settings.cpp \
     classes/simple_json.cpp \
     installer/install_blishhud.cpp \
     installer/install_boontable.cpp \
@@ -13,7 +14,6 @@ SOURCES += \
     main.cpp \
     main_window/sources/mainwindow.cpp \
     classes/fileinteractions.cpp \
-    classes/settings.cpp \
     classes/updater.cpp \
     updater/config/githup_config.cpp \
     updater/githupdater.cpp
@@ -22,6 +22,7 @@ HEADERS += \
     helpers/includes/downloader.h \
     helpers/includes/logger.h \
     helpers/includes/version.h \
+    helpers/includes/settings.h \
     classes/simple_json.h \
     installer/install_blishhud.h \
     installer/install_boontable.h \
@@ -32,7 +33,6 @@ HEADERS += \
     commandstring_editor/includes/form.h \
     main_window/includes/mainwindow.h \
     classes/fileinteractions.h \
-    classes/settings.h \
     classes/updater.h \
     updater/config/githup_config.h \
     updater/githupdater.h
